backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.0
email-validator==2.2.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
